admind view for users and places

+++ project/services/place.service.server.js
@@ -3,6 +3,7 @@ module.exports = function (app, model) {
   var placeModel = model.placeModel;
 
   app.post('/project/api/places', createPlace);
+  app.get('/project/api/places', findAllPlaces);
   app.get('/project/api/places/:placeId', findPlaceById);
   app.get('/project/api/places/google/:googleId', findPlaceByGoogleId);
   app.post('/project/api/places/:placeId/reviews', addPlaceReview);
@@ -21,6 +22,14 @@ module.exports = function (app, model) {
     } else {
       res.status(400).send('Invalid request body.');
     }
+  }
+
+  function findAllPlaces(req, res) {
+    placeModel.findAllPlaces().then(function (places) {
+      res.json(places);
+    }, function (err) {
+      res.sendStatus(500);
+    });
   }
 
   function findPlaceById(req, res) {
@@ -68,21 +77,29 @@ module.exports = function (app, model) {
   function updatePlace(req, res) {
     var place = req.body;
     if (place) {
-      placeModel.updatePlace(req.params.placeId, place).then(function () {
-        res.sendStatus(200);
-      }, function () {
-        res.sendStatus(500);
-      });
+      if (req.user.role === 'ADMIN') {
+        placeModel.updatePlace(req.params.placeId, place).then(function () {
+          res.sendStatus(200);
+        }, function () {
+          res.sendStatus(500);
+        });
+      } else {
+        res.status(409).send('Yod do not have permission to do that.');
+      }
     } else {
       res.status(400).send('Invalid request body.');
     }
   }
 
   function deletePlace(req, res) {
-    placeModel.deletePlace(req.params.placeId).then(function () {
-      res.sendStatus(200);
-    }, function (error) {
-      res.sendStatus(500);
-    });
+    if (req.user.role === 'ADMIN') {
+      placeModel.deletePlace(req.params.placeId).then(function () {
+        res.sendStatus(200);
+      }, function (error) {
+        res.sendStatus(500);
+      });
+    } else {
+      res.status(409).send('Yod do not have permission to do that.');
+    }
   }
 };
